SQUASH => Create address prefix map...

+++ common/types.go
@@ -648,5 +648,5 @@ func (l Location) ContainsAddress(a Address) bool {
 		log.Fatal("unable to get address prefix range for location")
 	}
 	// Ranges are fully inclusive
-	return prefix >= prefixRange.lo && prefix <= prefixRange.hi
+	return uint8(prefix) >= prefixRange.lo && uint8(prefix) <= prefixRange.hi
 }
